Fix bug where multiple requests called
Bug was caused by the combination of polling and changing of queryData. When the inner observable (queryData) was updated, it was called the same number of times that the outer observable (polling timer) had values. Fixed by only polling on the last value of the inner observable (queryData) and cancelling the polling obserable and creating a new one whenever a new value of queryData is given